+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 ### Removed
 - Removed generation of base/core files.
 - Removed `-module` option from `pywinrt.exe`.
+- Removed `winrt.system.Guid` type alias.
 
 ## [v1.0.0-beta.10] - 2023-08-11